requirements.txt
feedparser
jinja2
markdown
pyyaml
requests
websocket-client
youtube-transcript-api
yt-dlp
